Add static camera and empty hub player controller
This commit adds a static camera to the hub scene and I've added an empty player controller that will have different movement and control scheme that way I can make the scene closer to an "animal crossing" view and controls. I also added colliders to the floor and desk. I need to add colliders to the door, then add walls, so the player doesn't walk off the map