add interaction ability to door and laptop
The door and laptop now can be interacted with the E/Left Mouse Click/Gamepad A input. I tried to implement the highlighting but I don't know how to work with meshes, so I decided to leave it out. I will most likely replace this with a UI prompt instead. The door when interacted with quits the game. The laptop when interacted with does nothing